It is first commit for this repository
All the work was done for the sake of expanding knowledge in the field of programming related to the Java programming language and the JavaFX library. All the work was done on the instructions of the university in the spring semester of 2022. There's nothing personal about it. And I don't take any responsibility.

To understand this code, you need to know about Java Core and the JavaFX library.